+++ Assignment3/src/academy/pocu/comp2500/assignment3/Turret.java
@@ -9,8 +9,8 @@ public final class Turret extends Unit implements IThinkable {
     private static final UnitType UNIT_TYPE = UnitType.GROUND;
     private static final byte VISION = 2;
     private static final byte AOE = 0;
-    private static final byte MAX_HP = 99;
-    private static final byte AP = 7;
+    private static final int MAX_HP = 99;
+    private static final int AP = 7;
     private static final List<UnitType> TARGETABLE = Arrays.asList(UnitType.AIR);
 
     public Turret(IntVector2D position) {
